+++ woocommerce-mastercard.php
@@ -1,14 +1,14 @@
 <?php
 /**
  * Plugin Name: Mastercard Payment Gateway Services
- * Description: Accept payments on your WooCommerce store using the Mastercard Payment Gateway Services. Requires PHP 7.4+ & WooCommerce 7.3+
+ * Description: Accept payments on your WooCommerce store using the Mastercard Payment Gateway Services. Requires PHP 8.1+ & WooCommerce 7.3+
  * Plugin URI: https://github.com/fingent-corp/gateway-woocommerce-mastercard-module/
  * Author: Fingent Global Solutions Pvt. Ltd.
  * Author URI: https://www.fingent.com/
  * Tags: payment, payment-gateway, mastercard, mastercard-payements, mastercard-gateway, woocommerce-plugin, woocommerce-payment, woocommerce-extension, woocommerce-shop, mastercard, woocommerce-api
- * Version: 1.4.3
+ * Version: 1.4.4
  * Requires at least: 6.0
- * Tested up to: 6.4.2
+ * Tested up to: 6.5
  * Requires PHP: 7.4
  * php version 8.1
  *
@@ -16,7 +16,7 @@
  * WC tested up to: 8.5.2
  *
  * @package  Mastercard
- * @version  GIT: @1.4.3@
+ * @version  GIT: @1.4.4@
  * @link     https://github.com/fingent-corp/gateway-woocommerce-mastercard-module/
  */
 
@@ -37,14 +37,17 @@
  */
 
 if ( ! defined( 'ABSPATH' ) ) {
-	exit();
+	exit; // Exit if accessed directly.
 }
+
+use Automattic\WooCommerce\Utilities\OrderUtil;
+use Automattic\WooCommerce\Internal\Features\FeaturesController;
 
 /**
  * Main class of the Mastercard Payment Gateway Services Module
  *
  * @package  Mastercard
- * @version  Release: @1.4.3@
+ * @version  Release: @1.4.4@
  * @link     https://github.com/fingent-corp/gateway-woocommerce-mastercard-module/
  */
 class WC_Mastercard {
@@ -54,13 +57,28 @@ class WC_Mastercard {
 	 *
 	 * @var WC_Mastercard
 	 */
-	private static $instance;
+	private static $instance = null;
 
 	/**
 	 * WC_Mastercard constructor.
 	 */
 	public function __construct() {
 		add_action( 'plugins_loaded', array( $this, 'init' ) );
+		add_action( 'init', array( $this, 'clean_output_buffer' ) );
+		add_action( 'before_woocommerce_init', array( $this, 'declare_cart_checkout_blocks_compatibility' ) );
+		if ( ! $this->is_order_pay_page() ) {
+			add_action( 'woocommerce_blocks_loaded', array( $this, 'woocommerce_gateway_mastercard_woocommerce_block_support' ), 99 );
+		}
+	}
+
+	/**
+	 * Clear output bufffer.
+	 *
+	 * @version 1.0
+	 * @package Helpfie
+	 */
+	public function clean_output_buffer() {
+		ob_start();
 	}
 
 	/**
@@ -69,26 +87,23 @@ class WC_Mastercard {
 	 * @return void
 	 */
 	public function init() {
-
-		define( 'MPGS_PLUGIN_FILE', __FILE__ );
-		define( 'MPGS_PLUGIN_BASENAME', plugin_basename( MPGS_PLUGIN_FILE ) );
-
+		define( 'MPGS_TARGET_PLUGIN_FILE', __FILE__ );
+		define( 'MPGS_TARGET_PLUGIN_BASENAME', plugin_basename( MPGS_TARGET_PLUGIN_FILE ) );
 		add_action( 'admin_init', array( $this, 'stop' ) );
-
 		if ( ! class_exists( 'WC_Payment_Gateway' ) ) {
 			return;
 		}
-
 		define( 'MPGS_ISO3_COUNTRIES', include plugin_basename( '/iso3.php' ) );
 		require_once plugin_basename( '/vendor/autoload.php' );
 		require_once plugin_basename( '/includes/class-gateway.php' );
-
+		require_once plugin_basename( '/includes/class-gateway-notification.php' );
 		load_plugin_textdomain( 'mastercard', false, trailingslashit( dirname( plugin_basename( __FILE__ ) ) ) . 'i18n/' );
-
+		// phpcs:disable WordPress.Security.EscapeOutput.OutputNotEscaped, WordPress.Security.NonceVerification.Recommended
 		add_filter(
 			'woocommerce_order_actions',
 			function ( $actions ) {
-				$order_id = isset( $_REQUEST['post'] ) ? sanitize_text_field( wp_unslash( $_REQUEST['post'] ) ) : null;
+				$hpos_enabled = get_option( 'woocommerce_custom_orders_table_enabled' );
+				$order_id     = self::get_order_id();
 				if ( $order_id ) {
 					$order = new WC_Order( $order_id );
 					if ( $order->get_payment_method() === Mastercard_Gateway::ID ) {
@@ -99,15 +114,13 @@ class WC_Mastercard {
 						}
 					}
 				}
-
 				return $actions;
 			}
 		);
-
+		// phpcs:enable WordPress.Security.EscapeOutput.OutputNotEscaped, WordPress.Security.NonceVerification.Recommended
 		add_filter( 'woocommerce_payment_gateways', array( $this, 'add_gateways' ) );
 		add_filter( 'plugin_action_links_' . plugin_basename( __FILE__ ), array( $this, 'plugin_action_links' ) );
 		add_filter( 'plugin_row_meta', array( __CLASS__, 'plugin_row_meta' ), 10, 2 );
-
 		add_action(
 			'rest_api_init',
 			function () {
@@ -124,7 +137,7 @@ class WC_Mastercard {
 									return is_numeric( $param );
 								},
 							),
-						)
+						),
 					)
 				);
 				register_rest_route(
@@ -140,7 +153,7 @@ class WC_Mastercard {
 									return is_numeric( $param );
 								},
 							),
-						)
+						),
 					)
 				);
 				register_rest_route(
@@ -156,7 +169,7 @@ class WC_Mastercard {
 									return is_numeric( $param );
 								},
 							),
-						)
+						),
 					)
 				);
 				register_rest_route(
@@ -173,13 +186,21 @@ class WC_Mastercard {
 	}
 
 	/**
+	 * Get the plugin url.
+	 *
+	 * @return string
+	 */
+	public static function plugin_url() {
+		return untrailingslashit( plugins_url( '/', MPGS_TARGET_PLUGIN_FILE ) );
+	}
+
+	/**
 	 * Check if WooCommerce is active or not.
 	 *
 	 * @since 1.2.0
 	 *
 	 * @return bool
 	 */
-
 	private function woocommerce_is_active() {
 		return is_plugin_active( 'woocommerce/woocommerce.php' );
 	}
@@ -189,12 +210,10 @@ class WC_Mastercard {
 	 *
 	 * @since 1.2.0
 	 *
-	 * @return bool
+	 * @return void
 	 */
-
 	public function stop() {
 		if ( ! $this->woocommerce_is_active() ) {
-
 			deactivate_plugins( plugin_basename( __FILE__ ) );
 			unset( $_GET['activate'] );
 			add_action( 'admin_notices', array( $this, 'admin_notices' ) );
@@ -206,13 +225,12 @@ class WC_Mastercard {
 	 *
 	 * @since 1.2.0
 	 *
-	 * @return bool
+	 * @return void
 	 */
-
 	public function admin_notices() {
-		$class = 'notice notice-error';
+		$class   = 'notice notice-error';
 		$message = __( 'Kindly ensure the WooCommerce plugin is active before activating the Mastercard Payment Gateway Services plugin.', 'mastercard' );
-		printf( '<div class="%1$s"><p>%2$s</p></div>', esc_attr( $class ), esc_html( $message ) ); 
+		printf( '<div class="%1$s"><p>%2$s</p></div>', esc_attr( $class ), esc_html( $message ) );
 	}
 
 	/**
@@ -313,7 +331,7 @@ class WC_Mastercard {
 	 * @return array
 	 */
 	public static function plugin_row_meta( $links, $file ) {
-		if ( MPGS_PLUGIN_BASENAME !== $file ) {
+		if ( MPGS_TARGET_PLUGIN_BASENAME !== $file ) {
 			return $links;
 		}
 
@@ -333,10 +351,76 @@ class WC_Mastercard {
 
 		$row_meta = array(
 			'docs'    => '<a href="' . esc_url( $docs_url ) . '" aria-label="' . esc_attr__( 'View mastercard documentation', 'mastercard-payment-gateway-services' ) . '">' . esc_html__( 'Docs', 'mastercard-payment-gateway-services' ) . '</a>',
-			'support' => '<a href="' . esc_url( $support_url ) . '" aria-label="' . esc_attr__( 'Visit mastercard support', 'mastercard-payment-gateway-services' ) . '">' . esc_html__( 'Support', 'mastercard-payment-gateway-services' ) . '</a>'
+			'support' => '<a href="' . esc_url( $support_url ) . '" aria-label="' . esc_attr__( 'Visit mastercard support', 'mastercard-payment-gateway-services' ) . '">' . esc_html__( 'Support', 'mastercard-payment-gateway-services' ) . '</a>',
 		);
 
 		return array_merge( $links, $row_meta );
+	}
+
+	/**
+	 * Return WooCommerce order id.
+	 *
+	 * @return int Order id.
+	 */
+	public static function get_order_id() {
+		if ( 'yes' !== self::is_hpos() ) {
+			$order_id = isset( $_REQUEST['post'] ) ? sanitize_text_field( wp_unslash( $_REQUEST['post'] ) ) : null; // phpcs:ignore WordPress.Security.NonceVerification.Recommended
+		} else {
+			$order_id = isset( $_REQUEST['id'] ) ? sanitize_text_field( wp_unslash( $_REQUEST['id'] ) ) : null; // phpcs:ignore WordPress.Security.NonceVerification.Recommended
+		}
+
+		return $order_id;
+	}
+
+	/**
+	 * Confirm whether HPOS has been enabled or not.
+	 *
+	 * @return bool HPOS.
+	 */
+	public static function is_hpos() {
+		return OrderUtil::custom_orders_table_usage_is_enabled() ? 'yes' : 'no';
+	}
+
+	/**
+	 * Is_order_pay_page - Returns true when viewing the order received page.
+	 *
+	 * @return bool
+	 */
+	public function is_order_pay_page() {
+		return isset( $_GET['key'] ) && isset( $_SERVER['REQUEST_URI'] ) && strpos( wp_unslash( $_SERVER['REQUEST_URI'] ), 'order-pay' ) !== false; // phpcs:ignore WordPress.Security.NonceVerification.Recommended, WordPress.Security.ValidatedSanitizedInput.InputNotSanitized
+	}
+
+	/**
+	 * Function to declare compatibility with cart_checkout_blocks feature.
+	 *
+	 * @since 1.4.4
+	 * @return void
+	 */
+	public function declare_cart_checkout_blocks_compatibility() { // phpcs:ignore Universal.Files.SeparateFunctionsFromOO.Mixed
+		if ( class_exists( '\Automattic\WooCommerce\Utilities\FeaturesUtil' ) ) {
+			\Automattic\WooCommerce\Utilities\FeaturesUtil::declare_compatibility( 'cart_checkout_blocks', __FILE__, true );
+			\Automattic\WooCommerce\Utilities\FeaturesUtil::declare_compatibility( 'custom_order_tables', __FILE__, true );
+		}
+	}
+
+	/**
+	 * Function to register the Mastercard payment method type.
+	 *
+	 * @since 1.4.4
+	 * @return void
+	 */
+	public function woocommerce_gateway_mastercard_woocommerce_block_support() {
+		if ( ! class_exists( 'Automattic\WooCommerce\Blocks\Payments\Integrations\AbstractPaymentMethodType' ) ) {
+			return;
+		}
+
+		require_once plugin_dir_path( __FILE__ ) . 'includes/class-gateway-blocks-support.php';
+		add_action(
+			'woocommerce_blocks_payment_method_type_registration',
+			function ( Automattic\WooCommerce\Blocks\Payments\PaymentMethodRegistry $payment_method_registry ) {
+				$payment_method_registry->register( new Mastercard_Gateway_Blocks_Support() );
+			}
+		);
 	}
 }